% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_density.R
\name{CalculateDensity}
\alias{CalculateDensity}
\title{Compute Simulation Heatmaps}
\usage{
CalculateDensity(
  sim_object,
  steps = NULL,
  which = "all",
  step_size = 1,
  cores = 1
)
}
\arguments{
\item{sim_object}{object created with \code{CreateSimulationObject}}

\item{steps}{which simulation steps to compute heatmaps for (Tissue, Holes, Cells, or All)}

\item{which}{which simulation to compute it for}

\item{step_size}{resolution of heatmap}

\item{cores}{number of cpu cores}
}
\value{
a new \code{SpatSimObj} with probability densities calculated
}
\description{
Compute Simulation Heatmaps
}
