% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_spatial_processes.R
\name{GenerateSpatialPattern}
\alias{GenerateSpatialPattern}
\title{Generate Spatial Point Pattern}
\usage{
GenerateSpatialPattern(sim_object, lambda = 25, ...)
}
\arguments{
\item{sim_object}{A 'SpatSimObj' containing a window.}

\item{lambda}{The intensity of the point pattern Default is 25.}

\item{...}{Additional arguments passed to 'rpoispp'.}
}
\value{
The updated 'sim_object' with a simulated point process added to the 'Processes' slot.
}
\description{
Generate a spatial point pattern within the simulation object's window using a Poisson point process.
}
\details{
This function generates a spatial point process within the window of the 'sim_object'
using a Poisson point pattern with intensity 'lambda'. The simulated point pattern is added
to the 'Patterns' slot of the 'sim_object'. Additional arguments can be passed to the 'rpoispp' function.
}
\examples{
sim_object <- CreateSimulationObject()
sim_object <- GenerateSpatialPattern(sim_object, lambda = 30)
}
