% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_spatial_processes.R
\name{GenerateSpatialProcess}
\alias{GenerateSpatialProcess}
\title{Generate Spatial Point Process}
\usage{
GenerateSpatialProcess(sim_object, lambda = 25, ...)
}
\arguments{
\item{sim_object}{A 'Spatial Simulation Object' containing a window.}

\item{lambda}{The intensity of the point process. Default is 25.}

\item{...}{Additional arguments passed to 'rpoispp'.}
}
\value{
The updated 'sim_object' with a simulated point process added to the 'Processes' slot.
}
\description{
Generate a spatial point process within the simulation object's window using a Poisson point process.
}
\details{
This function generates a spatial point process within the window of the 'sim_object'
using a Poisson point process with intensity 'lambda'. The simulated point process is added
to the 'Processes' slot of the 'sim_object'. Additional arguments can be passed to the 'rpoispp' function.
}
\examples{
sim_object <- create_simObject()
sim_object <- GenerateSpatialProcess(sim_object, lambda = 30)
}
