% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot,Spatial Simulation Object-method}
\alias{plot,Spatial Simulation Object-method}
\title{Method \code{plot} for objects of class \verb{Spatial Simulation Object}}
\usage{
\\method{plot}{`Spatial Simulation Object`}
}
\arguments{
\item{x}{Spatial Simulation Object}

\item{nrow}{number of rows in the output plot}

\item{ncol}{number of columns in the output plot}

\item{which}{which spatial pattern(s) to plot}

\item{what}{currently only accepting "Processes" but potentially more things using the spatstat plot function in the future}
}
\description{
When proceeding though the simulation of multiplex data, it is nice to view some of the processes and results of the process.
Using the plot function, the \code{plot} from spatstat is used (in the case of plotting just the point process). When plotting
more detail from the \verb{Spatial Simulation Object}, please use \code{PlotSimulation}
}
