% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot.SpatialSimulationObject}
\alias{plot.SpatialSimulationObject}
\title{plot function for SpatialSimulationObject}
\usage{
\method{plot}{SpatialSimulationObject}(x, ...)
}
\arguments{
\item{x}{of class \code{SpatialSimulationObject}}

\item{...}{other things to pass to the plot method for SpatialSimulationObjects including \code{nrow}, \code{ncol} for the number of rows and columns of plots,
\code{which} processes to plot, and \code{what} which currently only works with "Processes" but may be updated in the future}
}
\description{
plot function for SpatialSimulationObject
}
