% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary.R
\name{SummariseSpatial}
\alias{SummariseSpatial}
\title{Summarise Spatial}
\usage{
SummariseSpatial(spatial_list, markers)
}
\arguments{
\item{spatial_list}{list of spatial data frames with \code{markers} column names}

\item{markers}{names of columns, probably cell types, that contain 1s and 0s representing positive/negative assignments}
}
\value{
data frome with summary counts and proportions for the markers in each spatial data frame
}
\description{
Summarise Spatial
}
