% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary,Spatial Simulation Object-method}
\alias{summary,Spatial Simulation Object-method}
\title{Method \code{summary} for objects in the \code{mIFsim} package}
\usage{
\\method{summary}{`Spatial Simulation Object`}
}
\arguments{
\item{object}{Spatial Simulation Object}
}
\description{
This provides an easy summary of what has already been executed in the Spatial Simulation Object.
Things like the number of simulations, number of cell types desired, as well as information about what has been simulated
are included in the sumamry output.
}
