% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.SpatialSimulationObject}
\alias{summary.SpatialSimulationObject}
\title{summary function for SpatialSimulationObject}
\usage{
\method{summary}{SpatialSimulationObject}(object, ...)
}
\arguments{
\item{object}{of class \code{SpatialSimulationObject}}

\item{...}{nothing else to pass to summary if object is a SpatialSimulationObject}
}
\description{
summary function for SpatialSimulationObject
}
