% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.SpatSimObj}
\alias{summary.SpatSimObj}
\title{summary function for SpatSimObj}
\usage{
\method{summary}{SpatSimObj}(object, ...)
}
\arguments{
\item{object}{of class \code{SpatSimObj}}

\item{...}{nothing else to pass to summary if object is a SpatSimObj}
}
\description{
summary function for SpatSimObj
}
