% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wm}
\alias{wm}
\title{Round spatstat window}
\format{
An \code{owin} object with many slots describing the size and mask
\describe{
\item{type}{mask}
}
}
\usage{
wm
}
\description{
A mask that turns a window from spatstat round to mimic a window that is similar to most tissue microarray image scans
}
\keyword{datasets}
